Improve export test utility and constant sharing
